+++ apps/e2e/tests/backend/endpoints/api/v1/integrations/neon/projects/provision.test.ts
@@ -25,7 +25,6 @@ it("should be able to provision a new project if neon client details are correct
       "headers": Headers { <some fields may have been hidden> },
     }
   `);
-  console.log(response.body);
 
 
   // test API keys
@@ -63,10 +62,12 @@ it("should be able to provision a new project if neon client details are correct
           "oauth_providers": [
             {
               "id": "github",
+              "provider_config_id": "github",
               "type": "shared",
             },
             {
               "id": "google",
+              "provider_config_id": "google",
               "type": "shared",
             },
           ],