+++ apps/e2e/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @stackframe/e2e-tests
 
+## 2.8.23
+
+### Patch Changes
+
+- Updated dependencies
+  - @stackframe/stack-shared@2.8.23
+  - @stackframe/js@2.8.23
+
+## 2.8.22
+
+### Patch Changes
+
+- Updated dependencies
+  - @stackframe/stack-shared@2.8.22
+  - @stackframe/js@2.8.22
+
 ## 2.8.21
 
 ### Patch Changes