+++ apps/mock-oauth-server/CHANGELOG.md
@@ -1,5 +1,9 @@
 # @stackframe/mock-oauth-server
 
+## 2.8.23
+
+## 2.8.22
+
 ## 2.8.21
 
 ## 2.8.20